error checking added

+++ repositories/entry.repository.ts
@@ -9,17 +9,45 @@ export default class EntryRepository {
 
   async delete(id: string) {
     const db = await dbPromise;
+
+    const entry = await db.get('entries', id);
+
+    if (!entry) {
+      throw new Error('Entry not found.');
+    }
+
     await db.delete('entries', id);
   }
 
-  async update(entry: Entry) {
+  async update(id: string, updates: Partial<Entry>) {
     const db = await dbPromise;
-    entry.updatedAt = Date.now();
-    await db.put('entries', entry);
+
+    const entry = await db.get('entries', id);
+
+    if (!entry) {
+      throw new Error('Entry not found.');
+    }
+
+    await db.put('entries', {
+      ...entry,
+      ...updates,
+      updatedAt: Date.now(),
+    });
   }
 
   async getChildren(parentId: string | null) {
     const db = await dbPromise;
+
+    if (!parentId) {
+      return [];
+    }
+
+    const entry = await db.get('entries', parentId);
+
+    if (!entry) {
+      throw new Error('Entry not found.');
+    }
+
     const entries = await db.getAll('entries');
     return entries.filter((entry) => entry.parentId === parentId);
   }
